LPD-42997 Move to a component
