+++ modules/apps/fragment/fragment-impl/build.gradle
@@ -25,6 +25,7 @@ dependencies {
 	compileOnly project(":apps:frontend-token:frontend-token-definition-api")
 	compileOnly project(":apps:info:info-api")
 	compileOnly project(":apps:item-selector:item-selector-api")
+	compileOnly project(":apps:item-selector:item-selector-criteria-api")
 	compileOnly project(":apps:layout:layout-api")
 	compileOnly project(":apps:layout:layout-display-page-api")
 	compileOnly project(":apps:layout:layout-page-template-api")
